Fix PR CI for project explorer fallback
