+++ src/components/Explorer/utils.ts
@@ -1,4 +1,8 @@
 import type { CustomIconName } from '@src/components/CustomIcon'
+import {
+  FILE_PLACEHOLDER_NAME,
+  FOLDER_PLACEHOLDER_NAME,
+} from '@src/components/Explorer/placeholders'
 import { sortFilesAndDirectories } from '@src/lib/desktopFS'
 import {
   desktopSafePathJoin,
@@ -10,6 +14,11 @@ import {
 import type { FileEntry } from '@src/lib/project'
 import type { SubmitByPressOrBlur } from '@src/lib/types'
 import type { ReactNode } from 'react'
+export {
+  addPlaceHoldersForNewFileAndFolder,
+  FILE_PLACEHOLDER_NAME,
+  FOLDER_PLACEHOLDER_NAME,
+} from '@src/components/Explorer/placeholders'
 
 /**
  * Remap FileEntry data into another data structure for the Project Explorer
@@ -170,42 +179,6 @@ export const flattenProject = (
     )
   }
   return flattenTreeInOrder
-}
-
-/**
- * In memory add fake placeholders for folder and files.
- * @param children
- * @param parentPath absolute path
- * @returns
- */
-export const addPlaceHoldersForNewFileAndFolder = (
-  children: FileEntry[] | null,
-  parentPath: string
-) => {
-  if (children === null) {
-    return
-  }
-
-  for (let i = 0; i < children.length; i++) {
-    addPlaceHoldersForNewFileAndFolder(
-      children[i].children,
-      joinOSPaths(parentPath, children[i].name)
-    )
-  }
-
-  const placeHolderFolderEntry: FileEntry = {
-    path: joinOSPaths(parentPath, FOLDER_PLACEHOLDER_NAME),
-    name: FOLDER_PLACEHOLDER_NAME,
-    children: [],
-  }
-  children.unshift(placeHolderFolderEntry)
-
-  const placeHolderFileEntry: FileEntry = {
-    path: joinOSPaths(parentPath, FILE_PLACEHOLDER_NAME),
-    name: FILE_PLACEHOLDER_NAME,
-    children: null,
-  }
-  children.push(placeHolderFileEntry)
 }
 
 export const isRowFake = (
@@ -409,8 +382,6 @@ export function shouldDroppedEntryBeMoved(
 export const NOTHING_IS_SELECTED: number = -2
 export const CONTAINER_IS_SELECTED: number = -1
 export const STARTING_INDEX_TO_SELECT: number = 0
-export const FOLDER_PLACEHOLDER_NAME = '.zoo-placeholder-folder'
-export const FILE_PLACEHOLDER_NAME = '.zoo-placeholder-file.kcl'
 
 /**
  * Check if a drag event contains external files (not internal drag)
